migrations i model chall

+++ web/trivia4-app/database/migrations/2022_11_28_123417_create_challenges_table.php
@@ -15,9 +15,9 @@ class CreateChallengesTable extends Migration
     {
         Schema::create('challenges', function (Blueprint $table) {
             $table->id();
-            $table->int('challenger');
-            $table->int('challenged');
-            $table->int('winner');
+            $table->unsignedBigInteger('challenger');
+            $table->unsignedBigInteger('challenged');
+            $table->unsignedBigInteger('winner');
 
         });
     }
